refactor: 로테이션 로직 리팩토링 #9
