+++ src/app/rotation/page.tsx
@@ -1,5 +1,5 @@
-import RotationListForAll from "@/components/_/RotationListForAll";
-import RotationListForNew from "@/components/_/RotationListForNew";
+import RotationListForAll from "@/components/_rotation/RotationListForAll";
+import RotationListForNew from "@/components/_rotation/RotationListForNew";
 
 const rotationPage = () => {
   return (
